% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_path.R
\name{add_path}
\alias{add_path}
\title{Add a New Path to the PATH Environment Variable}
\usage{
add_path(new_path)
}
\arguments{
\item{new_path}{The path that you want to add to the PATH.}
}
\value{
A character string of the new PATH environment variable.
}
\description{
This function takes a path as input and adds it to the current PATH environment variable.
}
\examples{
add_path("/usr/local/bin")
}
