% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close_all_sinks.R
\name{close_all_sinks}
\alias{close_all_sinks}
\title{Close All Sinks and Reset Message Output}
\usage{
close_all_sinks()
}
\value{
NULL invisibly.
}
\description{
This function closes all active sink connections and ensures that
subsequent outputs are displayed in the console.
}
\examples{
\dontrun{
  sink("output.txt")
  cat("This will be written to the file.")
  close_all_sinks()
  cat("This will be shown in the console.")
}
}
