% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_project_root.R
\name{find_project_root}
\alias{find_project_root}
\title{Find the Project Root}
\usage{
find_project_root(setwd = TRUE)
}
\arguments{
\item{setwd}{Logical. change the working directory to project root? Defaults to TRUE.}
}
\value{
The directory path containing an .Rproj file.
Throws an error if no .Rproj file is found in the directory tree above the current directory.
}
\description{
This function recursively searches for the project root, defined as the directory
containing an .Rproj file. It starts in the current directory and moves up the
directory tree until it either finds an .Rproj file or reaches a directory
with no parent.
}
\note{
Be careful when running this function, as it could end up searching a large
portion of your file system if no .Rproj file is found.
}
\examples{
\dontrun{
find_project_root()
}
}
