% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_run_num.R
\name{gen_run_num}
\alias{gen_run_num}
\title{Generate blocked non-NA run number variable using run-length encoding}
\usage{
gen_run_num(x)
}
\arguments{
\item{x}{vector of values with NAs interspersed}
}
\value{
blocked vector of "run number"
}
\description{
Generate blocked non-NA run number variable using run-length encoding
}
\author{
Nate Hall
}
