% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_todo.R
\name{github_todo}
\alias{github_todo}
\title{Update running TODO.md for display on github using \code{todor}}
\usage{
github_todo(...)
}
\description{
Pull all relevant todoR fields and render a TODO.md file that is viewable on github.
}
