% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llsync.R
\name{llsync}
\alias{llsync}
\title{llsync: Synchronize local and remote files using rsync}
\usage{
llsync(path_from, path_to, onyen = "natehall")
}
\arguments{
\item{path_from}{character, the source path of the file(s) to be synced.}

\item{path_to}{character, the destination path where the file(s) will be synced.}

\item{onyen}{user's onyen. defaults to nate's.}
}
\value{
Prints the output of the rsync command.
}
\description{
This function syncs files between the local machine and a remote server,
specifically the Longleaf cluster at UNC. The synchronization is done using the
rsync command. The function checks if either the source or the destination path
starts with "/proj", and if so, it adds the remote server address before the path.
}
\examples{
llsync("/local/path/file", "/proj/remote/path/")
llsync("/proj/remote/path/file", "/local/path/")
}
