% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_heatmaps.R, R/model_heatmaps.R
\name{lmer_heatmaps}
\alias{lmer_heatmaps}
\title{create a heatmap of fixed effects for a series of lmer model objects}
\usage{
lmer_heatmaps(
  lmer_list,
  box_width = 0.5,
  standardize = TRUE,
  coef_str_replace = NULL,
  substitue_str = "",
  viridis_theme = TRUE,
  background_base = 0,
  ...
)

lmer_heatmaps(
  lmer_list,
  box_width = 0.5,
  standardize = TRUE,
  coef_str_replace = NULL,
  substitue_str = "",
  viridis_theme = TRUE,
  background_base = 0,
  ...
)
}
\description{
TODO: document.

TODO: documnet.
}
