% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu_pretty.R
\name{menu_pretty}
\alias{menu_pretty}
\title{Pretty Menu with Default Option}
\usage{
menu_pretty(choices, title = NULL, default = 1)
}
\arguments{
\item{choices}{A character vector containing the menu options.}

\item{title}{An optional character string specifying the title of the menu. Default is NULL.}

\item{default}{An integer specifying the default option to select if the user hits return without making a selection. Default is 1.}
}
\value{
The integer index of the selected option.
}
\description{
Displays a menu with a list of choices and allows the user to make a selection.
If the user hits return without making a selection, a default option is chosen.
}
\examples{
choices <- c("Option 1", "Option 2", "Option 3")
selected <- menu_pretty(choices, title = "Please select an option:", default = 2)
cat("You selected option", selected, "\n")

}
