% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tree.R
\name{print_tree}
\alias{print_tree}
\title{Print a Tree-Like Structure of a List}
\usage{
print_tree(x, indent = "", last = TRUE)
}
\arguments{
\item{x}{A list object to be printed as a tree structure.}

\item{indent}{A string used internally by the function to manage indentation.
It is not intended to be set by the user.}

\item{last}{A logical value used internally to indicate if the current element
is the last in its sublist. It is not intended to be set by the user.}
}
\description{
Recursively prints a tree-like structure of a list object with proper
indentation to visualize its hierarchy. This function aims to mimic the
output of `lobstr::tree` with full values shown and colorized output for
the list elements.
}
\examples{
config <- list(
  run_behav = TRUE,
  paths = list(
    raw = list(
      behav = "/path/to/behav",
      eye = "/path/to/eye"
    )
  )
)
print_tree(config)

}
