% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_list_numeric.R
\name{reorder_list_numeric}
\alias{reorder_list_numeric}
\title{Reorder Data Frames in List by Numeric Name}
\usage{
reorder_list_numeric(df_list)
}
\arguments{
\item{df_list}{A list of data frames.}
}
\value{
A list of data frames ordered by the numeric values extracted from their names.
}
\description{
This function takes a list of data frames and reorders them based on the
numeric values embedded within their names.
}
\examples{
df1 <- data.frame(x = 1:3)
df713 <- data.frame(y = 4:6)
df42 <- data.frame(z = 7:9)

my_list <- list("42" = df42, "713" = df713, "1" = df1)
reordered_list <- reorder_list_numeric(my_list)
}
