% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_objects.R
\name{list_data_objects}
\alias{list_data_objects}
\title{List Data Objects in the Current Environment}
\usage{
list_data_objects()
}
\value{
A character vector of the names of data objects in the current environment.
}
\description{
This function lists all data objects in the current R environment, excluding functions.
}
\examples{
\dontrun{
  # Create some example data objects
  x <- 1:10
  y <- data.frame(a = 1:5, b = 6:10)
  z <- "Hello, world!"

  # Print data objects
  print(list_data_objects())
}
}
